technology strategy UI also draws the tech card
